Remove model specs
